+++ src/service/userService.ts
@@ -1,10 +1,53 @@
 import { UpdatedUserGetDTO } from "../interface/DTO/user/UpdatedUserGetDTO";
+import { SocialCreateRequestDTO } from "./../interface/DTO/auth/SocialCreateDTO";
 import { UserGetDTO } from "../interface/DTO/user/UserGetDTO";
+import { randomInitialNickname } from "../module/randomInitialNickname";
+
 import { PrismaClientKnownRequestError, PrismaClientValidationError } from "@prisma/client/runtime";
 import { PrismaClient } from "@prisma/client";
 import { dateConvertString, stringConvertTime } from "../module/convert/convertTime";
 
 const prisma = new PrismaClient();
+
+const getUserByEmail = async (socialCreateRequestDTO: SocialCreateRequestDTO) => {
+  try {
+    const userByEmail = await prisma.user.findFirst({
+      where: {
+        AND: [{ email: socialCreateRequestDTO.email }, { provider: socialCreateRequestDTO.provider }],
+      },
+    });
+    return userByEmail;
+  } catch (error) {
+    console.log(error);
+    throw error;
+  }
+};
+
+const createUser = async (socialCreateRequestDTO: SocialCreateRequestDTO, refreshToken: string) => {
+  try {
+    const newUser = await prisma.user.create({
+      data: {
+        nickname: randomInitialNickname(),
+        social_id: socialCreateRequestDTO.socialId,
+        email: socialCreateRequestDTO.email,
+        provider: socialCreateRequestDTO.provider,
+        refreshToken: refreshToken,
+      },
+    });
+    if (!newUser) return null;
+    return newUser;
+  } catch (error) {
+    if (error instanceof PrismaClientKnownRequestError) {
+      if (error.code == "P2002") {
+        // 이미 생성한 유저
+        return `이미 생성된 유저입니다.`;
+      }
+    } else if (error instanceof PrismaClientValidationError) {
+      return `${error.message}`;
+    }
+    throw error;
+  }
+};
 
 const signUp = async (machineId: string, nickname: string) => {
   try {
@@ -112,6 +155,8 @@ const updateUserNickname = async (machineId: string, nickname: string) => {
 };
 
 const userService = {
+  getUserByEmail,
+  createUser,
   signUp,
   getUser,
   updateUserNickname,
